Remove unnecessary PWA implementation
