+++ ui/package.json
@@ -41,10 +41,6 @@
 		"react-icons": "^5.0.1",
 		"react-router-dom": "^6.22.1",
 		"vite": "^5.0.12",
-		"vite-plugin-pwa": "^0.19.7",
-		"workbox-core": "^7.0.0",
-		"workbox-precaching": "^7.0.0",
-		"workbox-window": "^7.0.0",
 		"zod": "^3.22.4"
 	},
 	"devDependencies": {
