+++ ui/vite.config.mts
@@ -1,28 +1,10 @@
 import { fileURLToPath } from 'node:url';
 import { defineConfig } from 'vitest/config';
 import react from '@vitejs/plugin-react';
-import { VitePWA } from 'vite-plugin-pwa';
 
 // https://vitejs.dev/config/
 export default defineConfig({
-	plugins: [
-		react(),
-
-		VitePWA({
-			registerType: 'autoUpdate',
-			strategies: 'injectManifest',
-			srcDir: 'src',
-			filename: 'service-worker/implementation.ts',
-			manifest: {
-				name: 'scaledgitapp',
-				short_name: 'scaledgitapp',
-			},
-			devOptions: {
-				enabled: true,
-				type: 'module',
-			},
-		}),
-	],
+	plugins: [react()],
 	build: {
 		outDir: '../Server/wwwroot',
 		emptyOutDir: true,
